Make the utility class accept a DatasetContext instead of a MapReduceContext. In order to do this, it has to return the PartitionedFileSet, instead of setting it as input on the dataset.
